+++ TabLayout/TabGroup.js
@@ -167,11 +167,13 @@ const TabGroupBase = kind({
 		const children = React.useMemo(() => tabs.map(tab => {
 			if (tab) {
 				// eslint-disable-next-line no-shadow
-				const {icon, title, ...rest} = tab;
+				const {icon, title, tabKey, ...rest} = tab;
+				const key = tabKey || tabKey === 0 ? tabKey : `tabs_${title + (typeof icon === 'string' ? icon : '')}`;
+
 				return {
-					key: `tabs_${title + (typeof icon === 'string' ? icon : '')}`,
 					children: title,
 					icon,
+					key,
 					onFocusTab,
 					...rest
 				};